Boke out and renames a load more stuff in readiness for component/image subclassing.
